requirements.txt
scapy>=2.5.0
requests>=2.31.0
paho-mqtt>=1.6.1
rich>=13.7.0
aiofiles>=23.2.1
